schedule-bot/requirements.txt
python-telegram-bot[job-queue]==20.7
deep-translator==1.11.4
geopy==2.4.1
timezonefinder==6.2.0
